requirements.txt
streamlit==1.32.0
pandas==2.0.3
numpy==1.24.3
plotly==5.14.1
requests==2.28.1
python-dotenv==1.0.0
cryptography==41.0.3
python-nexo==1.0.0
yfinance==0.2.18
scipy==1.11.1
matplotlib==3.7.1
seaborn==0.12.2